fix: move FlavorsHome into separate module
