+++ UPDATE.md
@@ -278,3 +278,4 @@
 - 2025-10-28: Applied split gradient background from white to light gray for natural page sectioning.
 - 2025-10-28: Fixed redirect loop by sending unauthenticated users to the sign-in page.
 - 2025-10-29: Added settings toggle to enable or disable LLM logs with code 'cake2025'.
+- 2025-10-29: Moved FlavorsHome component into its own module to satisfy Next.js page export requirements.
